DOC (squash 3)
(1) DOC: Update conf.py

Add replace:: to epilog for link to intersphinx extension page

(2) DOC: Add sphobjinv.py docstrings

Add docstrings for encode & decode. May want to trim them down a bit, especially if some of the info is going to be elsewhere in the docs. OTOH, might be useful to have it in there explicitly.

(3) DOC: Prune docstrings, add notes for index

No need to overweight the docstrings with something easily enough
 put onto the main page.