DEV: Fix suppression of fuzzywuzzy speed warning
Warning is raised on *import*, not use, of fuzzywuzzy.process.extract.
Putting the warning suppression in the right place thus actually does
suppress the warning.

Also: add test to ensure that only the python-Levenshtein warning
is suppressed; complain if more than one warning is raised, or if
one warning is raised but it doesn't talk about Levenshtein.

Closes #24.