DEV: Add struct_dict schema
Believe struct_dict schema is now in place. Does not proofread the
value of the 'count' property, but seems to capture the rest of the
structure properly.

Does not proofread that the domain, role, or name values are sane.

Schema includes a 'metadata' property at the root level, to
satisfy #36.