DEV: Implement URL inventory import
Direct pulldown of inventories by URL now works, per #34.

Still needs testing added, which will require some augmentation
of the testing CLI, and decisions about what to test or not in
the CI. Also... should I post/host a static copy of the ref
inv somewhere...and if so, in both compressed and decompressed
form?

Probably.

[skip ci] since SourceTypes iteration test is known-broken.