DEV/TEST: Data file gen done, tested. BUGFIX!
BUGFIX: .zlib.compress was only picking up the first object
in the file due to a logic error in attempting to use a generator.
Now fixed.

DEV: Implement export of Inventory contents to a data file. TESTS
written that confirm original file imports the same content as
the exported/re-imported file.

TEST: Implement test cl arg to choose whether to bigtest all .inv's
in test/resource, or just attrs.

Nearly takes care of #53, just a test of two left.