DOC: Complete customfile; various other
Final draft of customfile.rst is done. Still should be looked
at a few more times, to iron out any last glitches.

Makefile is revised to allow for new 'sphmake' bash alias,
enabling passing of options to sphinx-build

Change doctest_global_setup code block so that _clear_files
explicitly works in the 'scratch' directory. The way it was before,
if a doctest happened to be written so that the working directory
changed, _clear_files would wreak messy havoc.