DOC: Complete main body of docs
Main portions of docs are now complete, for the project as it stands.

Complete levenshtein.rst, including chart, commentary, etc.

Fix bad cross-ref to 'evolve' (needed to point to SuperDataObj,
not DataObjStr) in api_usage.rst.

Change formatting of |soi| replace:: in conf.py, to bold-fixed,
to be less confusing. Other things formatted with the white
bounding box will generally be linked to something, whereas
"sphobjinv" won't. Having the formatting be different will
be clearer that the behavior is different.

Closes #47. Closes #38. Closes #23.

Still need to address README (#78) and 'intersphinx-xlswf'-as-
example-non-Sphinx-docset (#10), as well as general GH/etc.
descriptions, taglines, etc. (#30).