ADMIN/TEST: Complete README draft, robustify tests
For unclear reasons, on a different PC the regex engine behaves
differently, leading to a shell test failure that was working fine
when developed previously.

So, the shell tests blocks are now explicitly closed with a
reST comment.

The error message on a shell test failure is now clearer, also.

The README is fully drafted, I think.  Just proofing/revising
from here. #30 is nearly done; just need to look closely at the
ancillary copy for the project, make sure it's all up to date,
before actually closing the issue.

Add Travis 'install' command to create doc/scratch, to pave the
way for the CLI convert example in README.