Add Travis flake8 ver print; update for f8-2020
Codes changed in flake8-2020; what used to be a lump in YTT101
got split out; the particular code raised is now YTT303.

Helpful to print the `flake8 --version` invocation to see
the installed plugins.

Also made it so that the failed assert on a code search
prints the whole flake8 report, to allow seeing the dang thing.