Add POC linkcheck.py
Very simplistic, but works to report links in files. No actual
success/failure return code as yet.

Regex for URLs is ~probably~ terrible.

Redirects will be a headache.

Will want to be able to indicate known-broken links somehow.

Would be really nice to be able to reuse sphinx's linkcheck
tool, but not sure how to point it at an arbitrary file,
and especially an arbitrary non-RST file.

[skip ci]