Add special cases for OpenCV and celery
Celery has some duplicate objects in the inventory, which Sphinx
ignores but sphobjinv does not.

OpenCV has objects that are not parsed correctly by Sphinx after the
regex change in sphinx-doc/sphinx#8225, but that sphobjinv
currently reads without difficulty.

A decision will have to be made whether to follow sphinx#8225's fix for
the DOS potential reported in sphinx#8175, or implement one of
the other less-susceptible regexes suggested in sphinx#8175.

Prefer the latter option, as it won't require post-processing
of a combined domain:role.