Fix scratch-path fixture test
Definitely better to explicitly use {Enum}.value to get the
string value out.

This 3.11 change will probably have a similar knock-on effect
on pent... may or may not be able to use the shortcut I've
been using there for folding enum values into the minilanguage.

But, if I switch to an OO model anyways, it's ~moot.