REFACTOR: Better names, leaner code, new test
Rename a whole bunch of functions and parameters to make them
more intuitive.

Remove an interstitial function in the base-URL extraction code
that was simple enough to just be folded into an enclosing
function.

Add a test looking at base URL matching for a locally stored
inventory in tests/resource

[skip ci]