Add Azure Pipelines job to test sdist testability
(1) Adding to a temporary first stage for now, for faster testing.

Later to be moved to the aux tests at the end of the pipeline.

(2) Remove 'soi' dir & change to correct dir for steps

Helps to have cwd in the right place for things to, y'know, work.

(3) Escape those dollar signs

(4) Remove up-front docs build & req'ts install

For a build, it *should* be enough just to install 'build'?

(5) Unescape $s and turn 's to "s

Nope, wasn't the dollar signs.

Something weird in how the single quotes were interacting with
Pipelines's string handling, maybe?

(6) Add a debugging step

(7) Refine the regex for finding the sdist dir