Add Python 3.12, retire Python 3.7
Assuming we want to support four active Python versions, retire Python
3.7, which is end-of-life since nearly a year ago, and add support for
Python 3.12.

New Python versions are released once a year, and make for a good
reason to make a release to PyPI, even if there haven't been any major
changes to the code base. Not making a release for a year results in
warnings such as "An important project maintenance signal to consider
for sphobjinv is that it hasn't seen any new versions released to PyPI
in the past 12 months, and could be considered as a discontinued
project, or that which receives low attention from its maintainers" on
https://snyk.io/advisor/python/sphobjinv, which companies use to
determine if a project is an acceptable dependency.