Convert textconv parser's 'version' report to 'version' action
Also make a short version output for textconf entrypoint.

Want 'infile' to show as required in the help, and looks like it's not
currently possible to render a multiline version string using
the 'version' action; and, treating -v as a normal flag means that
the inputs fail to parse when no 'infile' is passed.